added Query interface
