+++ src/main/java/com/spbgti/dispatcherapp/Entity/Event/ReadEvent.java
@@ -1,9 +1,9 @@
 package com.spbgti.dispatcherapp.Entity.Event;
 
 import com.spbgti.dispatcherapp.Entity.Event.Command.Query;
+import com.spbgti.dispatcherapp.Entity.Event.Command.QueryImpl;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import javax.persistence.EntityManager;
 import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.List;
@@ -24,7 +24,7 @@ public class ReadEvent extends Event {
     public List<Object> apply() throws ClassNotFoundException, NoSuchMethodException, InstantiationException, IllegalAccessException, InvocationTargetException {
         List<Object> result = new ArrayList<>();
         for (Query query : queries) {
-            result.add(query.apply());
+            result.add(query.execute());
         }
         return result;
     }
